Work on organization/rationale for setup
